moved middleware from api folder, split app and server logic
